feat: export dropdown is now on the notemetadataheader
