+++ src/views/note/NoteMetadataHeader.tsx
@@ -1,14 +1,119 @@
 import { ArrowPathIcon, DocumentDuplicateIcon, ShareIcon, TrashIcon } from '@heroicons/react/24/solid'
+import MoreVertIcon from '@mui/icons-material/MoreVert'
+import IconButton from '@mui/material/IconButton'
+import MenuItem from '@mui/material/MenuItem'
+import axios from 'axios'
 import moment from 'moment'
 import React from 'react'
 import { useNavigate } from 'react-router-dom'
 
 import LoaderIcon from '../../components/icons/LoaderIcon'
-import { useAuthContext } from '../../contexts/auth'
+import { StyledMenu } from '../../components/Menu/StyledMenu'
+import { TAuthContext, useAuthContext } from '../../contexts/auth'
 import { useGetAccount } from '../../hooks/api/accounts'
 import { useDeleteNoteInCurrentGroup, useGetNoteInCurrentGroup } from '../../hooks/api/notes'
-import { useNoteIdFromUrl } from '../../hooks/url'
+import { useGroupIdFromUrl, useNoteIdFromUrl} from '../../hooks/url'
 import { FormatMessage } from '../../i18n/TextComponent'
+import { API_BASE } from '../../lib/env'
+
+interface Props {
+  noteId: string 
+  groupId: string 
+}
+
+const extensions = ['md', 'pdf']
+
+export const NotesOptions = ( { noteId, groupId } : Props ) => {
+  const url = `${API_BASE}/groups/${encodeURIComponent(groupId)}/notes/${encodeURIComponent(noteId)}/export`
+  const auth : TAuthContext = useAuthContext()
+  const [anchorEl, setAnchorEl] = React.useState<null | HTMLElement>(null)
+  const open = Boolean(anchorEl)
+  const handleClick = (event: React.MouseEvent<HTMLButtonElement>) => {
+    setAnchorEl(event.currentTarget)
+  }
+  const handleClose = () => {
+    setAnchorEl(null)
+  }
+  
+  const handleExport = async (format: string) => {  
+    try {
+      const ids = [noteId, groupId]
+      const token = await auth.token()
+      
+      if (ids.some(id => !id)) {
+        throw new Error('ID is not defined')
+      }
+      if (!format || !extensions.some(ext => format == ext)) {
+        throw new Error('Invalid export format')
+      }
+      const response = await axios.get(url, {
+        headers: {
+          'Authorization': `Bearer ${token}`
+        },
+        responseType: 'blob',
+        params : {
+          format: format
+        }
+      })
+
+      // 😳 https://stackoverflow.com/questions/50694881/how-to-download-file-in-react-js
+      // create file link in browser's memory
+      const href = URL.createObjectURL(response.data)
+
+      // create "a" HTML element with href to file & click
+      const link = document.createElement('a')
+      link.href = href
+      link.setAttribute('download', `note.${format}`) //or any other extension
+      document.body.appendChild(link)
+      link.click()
+
+      // clean up "a" element & remove ObjectURL
+      document.body.removeChild(link)
+      URL.revokeObjectURL(href)
+
+    } catch (error) {
+      console.error(error)
+    }
+  }
+
+  return (
+    <div>
+      <IconButton
+        aria-label='more'
+        id='long-button'
+        aria-controls={open ? 'long-menu' : undefined}
+        aria-expanded={open ? 'true' : undefined}
+        aria-haspopup='true'
+        onClick={handleClick}
+      >
+        <MoreVertIcon />
+      </IconButton>
+      <StyledMenu
+        id='demo-customized-menu'
+        MenuListProps={{
+          'aria-labelledby': 'demo-customized-button',
+        }}
+        anchorEl={anchorEl}
+        open={open}
+        onClose={handleClose}
+      >
+        <MenuItem
+          onClick={() => {
+            handleExport('pdf')
+            handleClose()}}
+        >
+          <FormatMessage id='NOTE.export.button1' />
+        </MenuItem>
+        <MenuItem
+          onClick={() => {
+            handleExport('md')
+            handleClose()}}>
+          <FormatMessage id='NOTE.export.button2' />
+        </MenuItem>
+      </StyledMenu>
+    </div>
+  )
+}
 
 const NoteViewMetadataHeader: React.FC = () => {
   const navigate = useNavigate()
@@ -34,7 +139,7 @@ const NoteViewMetadataHeader: React.FC = () => {
       </div>
     </div>
     {/* Menu */}
-    <div className='grid grid-cols-[auto_auto_auto] gap-2'>
+    <div className='grid grid-cols-[auto_auto_auto_auto] gap-2'>
       {/* Duplicate */}
       <div className='group flex cursor-pointer items-center rounded p-1 hover:bg-blue-50' onClick={() => alert('Not Implemented')}>
         <DocumentDuplicateIcon className='mr-1 h-4 w-4 text-gray-500 group-hover:text-blue-500' />
@@ -56,6 +161,18 @@ const NoteViewMetadataHeader: React.FC = () => {
           <FormatMessage id='NOTE.delete' />
         </h5>
       </div>}
+      {/* BALISE */}
+      <div className='group flex cursor-pointer items-center p-1'>
+        {window.location.pathname.includes('/note/') &&
+            window.location.pathname.split('/')[4] &&
+              <NotesOptions 
+                noteId={useNoteIdFromUrl()}
+                groupId={useGroupIdFromUrl()}
+              />
+        }
+
+      </div>
+      {/* BALISE */}
     </div>
   </div>
 }
